requirements.txt
asttokens==2.4.1
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
executing==2.0.1
filelock==3.13.1
Flask==3.0.1
fonttools==4.47.0
fsspec==2023.12.2
idna==3.6
imutils==0.5.4
ipykernel==6.28.0
ipython==8.19.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jupyter_client==8.6.0
jupyter_core==5.7.0
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.5.8
networkx==3.2.1
numpy==1.26.3
opencv-python==4.9.0.80
packaging==23.2
parso==0.8.3
pillow==10.2.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
psutil==5.9.7
pure-eval==0.2.2
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pywin32==306
pyzmq==25.1.2
requests==2.31.0
scipy==1.11.4
six==1.16.0
stack-data==0.6.3
sympy==1.12
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
tornado==6.4
traitlets==5.14.1
typing_extensions==4.9.0
urllib3==2.1.0
wcwidth==0.2.13
Werkzeug==3.0.1
